refactor: move config file to upper level
