+++ src/server/adapters/index.ts
@@ -1,0 +1,2 @@
+export * from "./auth-prisma.ts";
+export * from "./auth-organization-prisma.ts";
